MM-13021
Encapsulate all xmlrpc related logic to the MAILAPI_Call class, specifically
the xmlrpc decoding, which otherwise needed to occur within MAILAPI_Client.

Our ideal here is that a client of the API doesn't need to understand the
transport mechanism for interacting with the service, abstracting this away
gets us closer to our goal.